+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ensure `Organization` CR in Azure MCs have the `subscriptionid` annotation set.
+
 ## [0.9.0] - 2021-07-05
 
 ### Added